Telephony: Fix issue of APN restore dialog displayed forever.
When DATA_CONNECTION_STATE_CHANGED intent is received with CONNECTED
state, show dialog is called if mRestoreDefaultApnMode == true.
This is creating two progress dialogs and when apn restore operation
is completed, only one dialog is dismissed.
To fix this, do not display restore apn dialog when receving intent
when mRestoreDefaultApnMode == true. The flag is set means that
the dialog is already being displayed.

Change-Id: I5c5aa0faa58e114cd6dee7cc34ac4e7c9a1389ed
CRs-Fixed: 1007758